move primitives out of cesiumHelpers
