+++ libraries/mapping/engines/cesium/src/lib/utils/primitives.ts
@@ -1,0 +1,27 @@
+import type { Primitive, Scene } from "cesium";
+
+export function getPrimitiveById(scene: Scene, id: string) {
+  const primitives = scene.primitives;
+  const length = primitives.length;
+
+  for (let i = 0; i < length; ++i) {
+    const p = primitives.get(i);
+    if (p.id === id) {
+      return p;
+    }
+  }
+
+  return null;
+}
+
+export function getAllPrimitives(scene: Scene) {
+  const primitives = scene.primitives;
+  const length = primitives.length;
+
+  const primitiveArray: Primitive[] = [];
+  for (let i = 0; i < length; ++i) {
+    const p = primitives.get(i);
+    primitiveArray.push(p);
+  }
+  return primitiveArray;
+}
